Modificacion en insertar albaranes

+++ SincronizadorQvetSuvesaPOS/SincronizadorQvetSuvesaPOS/Datos/Albaran_Detalle.cs
@@ -17,10 +17,10 @@ namespace SincronizadorQvetSuvesaPOS.Datos
         public long id { get; set; }
         public long idEncabezado { get; set; }
         public string Descripcion { get; set; }
-        public long CodigoInternoQvet { get; set; }
+        public int CodigoInternoQvet { get; set; }
         public decimal Cantidad { get; set; }
-        public long IVA { get; set; }
-        public double descuento { get; set; }
+        public int IVA { get; set; }
+        public decimal descuento { get; set; }
         public decimal PrecioVenta { get; set; }
         public decimal Total { get; set; }
         public string Unidad { get; set; }
